API Key 통신 방식 정립
v1/api/** -> JWT, APIKey 불필요
v2/api/** -> JWT만 필수
v3/api/** -> JWT 또는 APIKey 둘중 하나는 필수

+++ src/main/java/com/app/kokonut/auth/jwt/dto/AuthResponseDto.java
@@ -19,6 +19,6 @@ public class AuthResponseDto {
         private String grantType;
         private String accessToken;
         private String refreshToken;
-        private Integer refreshTokenExpirationTime;
+        private Long refreshTokenExpirationTime;
     }
 }
